requirements.txt
coc.py==2.0.1
discord.py==2.0.0
Flask==2.1.3
google-api-python-client==2.54.0
nest_asyncio==1.5.5
Pillow==9.2.0
requests
git+https://github.com/top-gg/python-sdk.git  # topggpy==2.0.0, not yet available with pip install topggpy == 2.0.0
